+++ packages/rum/src/domain/record/record.ts
@@ -21,8 +21,8 @@ import type { InputCallback } from './observers'
 import { getVisualViewport, getScrollX, getScrollY } from './viewports'
 import { assembleIncrementalSnapshot } from './utils'
 import { createElementsScrollPositions } from './elementsScrollPositions'
-import type { ShadowRootsController } from './shadowDom'
-import { initShadowRootsController } from './shadowDom'
+import type { ShadowRootsController } from './shadowRootsController'
+import { initShadowRootsController } from './shadowRootsController'
 
 export interface RecordOptions {
   emit?: (record: BrowserRecord) => void
