+++ packages/rum/src/domain/record/shadowRootsController.ts
